chore: Refactor SocietyWideImpact into new file
